requirements.txt
Flask==2.0.2
onnxruntime==1.10.0
Pillow==8.4.0
numpy==1.21.4